Added screen 3, changed screen 2
I coded out screen 3, and I removed ReviewPart1 and put its code into the designated screen (2_review_1)
